Fix ut for thinRuntime, To #43000843
