+++ pkg/ddc/thin/transform_fuse_test.go
@@ -17,9 +17,7 @@
 package thin
 
 import (
-	"fmt"
 	"reflect"
-	"sort"
 	"strings"
 	"testing"
 
@@ -736,47 +734,11 @@ func TestThinEngine_transformFuseWithDuplicateOptionKey(t1 *testing.T) {
 			t1.Errorf("transformFuse() error = %v", err)
 		}
 
-		value.Fuse.Envs = sortEnvVarByName(value.Fuse.Envs, common.ThinFuseOptionEnvKey)
+		value.Fuse.Envs = testutil.SortEnvVarByName(value.Fuse.Envs, common.ThinFuseOptionEnvKey)
 		if !testutil.DeepEqualIgnoringSliceOrder(t1, value.Fuse, wantValue.Fuse) {
 			valueYaml, _ := yaml.Marshal(value.Fuse)
 			wantYaml, _ := yaml.Marshal(wantValue.Fuse)
 			t1.Errorf("transformFuse() \ngot = %v, \nwant = %v", string(valueYaml), string(wantYaml))
 		}
 	})
-}
-
-// sortEnvVarByName sorts the value of an environment variable by its name
-func sortEnvVarByName(envs []corev1.EnvVar, name string) []corev1.EnvVar {
-	// Search for the environment variable with the given name
-	for i := range envs {
-		if envs[i].Name == name {
-			// Sort the value of the environment variable
-			envs[i].Value = sortEnvVarValue(envs[i].Value)
-			return envs
-		}
-	}
-	return envs
-}
-
-// sortEnvVarValue sorts the value of an environment variable
-func sortEnvVarValue(value string) string {
-	// Split the value into key=value pairs
-	pairs := strings.Split(value, ",")
-	kvMap := make(map[string]string, len(pairs))
-
-	// Add the key=value pairs to a map
-	for _, pair := range pairs {
-		kv := strings.Split(pair, "=")
-		if len(kv) == 2 {
-			kvMap[kv[0]] = kv[1]
-		}
-	}
-
-	// Sort the keys and build a new value
-	var sortedPairs []string
-	for key, value := range kvMap {
-		sortedPairs = append(sortedPairs, fmt.Sprintf("%s=%s", key, value))
-	}
-	sort.Strings(sortedPairs)
-	return strings.Join(sortedPairs, ",")
 }
